캐릭터 방향 및 애니메이션 코드 위치 변경 및 변수 추가/변경

+++ Monster.h
@@ -11,8 +11,8 @@ public:
 	bool IsCollide(int NewX, int NEwY);
 
 	
-	Uint32 ProcessTime;
-	Uint32 ElaspedTime;		// 경과 시간
+	Uint64 ProcessTime;
+	Uint64 ElaspedTime;		// 경과 시간
 };
 
 enum class EMonsterDirection
